+++ react/ProductSearchTemplate.js
@@ -9,15 +9,36 @@ import { SearchQueryContext } from './constants/searchContext'
 export default class SearchPage extends Component {
   static propTypes = {
     params: PropTypes.shape({
+      /** Brand name */
+      brand: PropTypes.string,
+
+      /** handles /:department/d
+       *  or /:department/:category
+       *  or /:department/:category/:subcategory */
+      department: PropTypes.string,
+      category: PropTypes.string,
+      subcategory: PropTypes.string,
+
       /** Search's term, e.g: eletronics. */
-      term: PropTypes.string.isRequired,
+      term: PropTypes.string,
     }),
     ...searchQueryPropTypes,
   }
 
   render() {
+    const props = {
+      ...this.props,
+      // todo: this logic should be in SearchQueryContainer
+      ...this.props.params.brand && {
+        query: {
+          ...this.props.query,
+          map: this.props.map || 'b',
+        }
+      }
+    }
+
     return (
-      <SearchQueryContainer {...this.props}>
+      <SearchQueryContainer {...props}>
         <SearchQueryContext.Consumer>
           {contextProps => <ExtensionPoint id="container" {...contextProps} />}
         </SearchQueryContext.Consumer>